feat: simplify project sidebar navigation
- Remove Kill Switches page (redundant — flags list already filters by type)
- Move Environments into Settings tab (CRUD + defaults in one unified table)
- Remove standalone Environments nav item
- Reorder sidebar: Flags, Lifecycle, Segments, Templates, Playground, Audit Log, Settings
- Environment default toggles now save immediately (no separate save button)
- Remove max-width constraint from project settings page
- Update E2E tests for new navigation structure
