+++ web/src/components/ProjectLayout.tsx
@@ -64,11 +64,9 @@ export default function ProjectLayout() {
       </div>
       <NavLink to={`/projects/${key}`} end className={navLinkClass} onClick={onNavigate}>Flags</NavLink>
       <NavLink to={`/projects/${key}/lifecycle`} className={navLinkClass} onClick={onNavigate}>Lifecycle</NavLink>
-      <NavLink to={`/projects/${key}/kill-switches`} className={navLinkClass} onClick={onNavigate}>Kill Switches</NavLink>
-      <NavLink to={`/projects/${key}/environments`} className={navLinkClass} onClick={onNavigate}>Environments</NavLink>
       <NavLink to={`/projects/${key}/segments`} className={navLinkClass} onClick={onNavigate}>Segments</NavLink>
-      <NavLink to={`/projects/${key}/playground`} className={navLinkClass} onClick={onNavigate}>Playground</NavLink>
       {canManageTemplates && <NavLink to={`/projects/${key}/templates`} className={navLinkClass} onClick={onNavigate}>Templates</NavLink>}
+      <NavLink to={`/projects/${key}/playground`} className={navLinkClass} onClick={onNavigate}>Playground</NavLink>
       <NavLink to={`/projects/${key}/audit-log`} className={navLinkClass} onClick={onNavigate}>Audit Log</NavLink>
       {isProjectAdmin && <NavLink to={`/projects/${key}/settings`} className={navLinkClass} onClick={onNavigate}>Settings</NavLink>}
     </>
